Refactor authentication components to use new ButtonSubmit and InputError components; remove deprecated SubmitButton and input error handling.
