+++ components/auth/register-form.tsx
@@ -6,8 +6,9 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { registerCredentials } from "@/actions/auth.action";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import SubmitButton from "@/components/stocks/submit-button";
 import { useActionState } from "react";
+import InputError from "@/components/stocks/input-error";
+import ButtonSubmit from "@/components/stocks/button-submit";
 
 export function RegisterForm() {
   const [state, formAction] = useActionState(registerCredentials, null);
@@ -27,9 +28,7 @@ export function RegisterForm() {
             <div className="grid gap-2">
               <Label htmlFor="name">Name</Label>
               <Input id="name" name="name" type="text" placeholder="John Doe" />
-              {state?.error?.name && (
-                <Label className="text-destructive">{state.error.name}</Label>
-              )}
+              <InputError message={state?.error?.name} />
             </div>
             <div className="grid gap-2">
               <Label htmlFor="email">Email</Label>
@@ -39,18 +38,12 @@ export function RegisterForm() {
                 type="email"
                 placeholder="m@example.com"
               />
-              {state?.error?.email && (
-                <Label className="text-destructive">{state.error.email}</Label>
-              )}
+              <InputError message={state?.error?.email} />
             </div>
             <div className="grid gap-2">
               <Label htmlFor="password">Password</Label>
               <Input id="password" name="password" type="password" />
-              {state?.error?.password && (
-                <Label className="text-destructive">
-                  {state.error.password}
-                </Label>
-              )}
+              <InputError message={state?.error?.password} />
             </div>
             <div className="grid gap-2">
               <Label htmlFor="confirmPassword">Confirm Password</Label>
@@ -59,13 +52,9 @@ export function RegisterForm() {
                 name="confirmPassword"
                 type="password"
               />
-              {state?.error?.confirmPassword && (
-                <Label className="text-destructive">
-                  {state.error.confirmPassword}
-                </Label>
-              )}
+              <InputError message={state?.error?.confirmPassword} />
             </div>
-            <SubmitButton submitting={"Registering"} submit={"Register"} />
+            <ButtonSubmit submitting={"Registering"} submit={"Register"} />
           </div>
         </div>
       </form>
